requirements.txt
python-dotenv
telebot
gspread
oauth2client
flask
